refactor: TokenAmount and its helper functions
